Upgrade Arrow & MobilityDB in the geo-cnpg image
Add Apache Arrow v19.0.0 (which requires ninja-build to build) to the
geo-cnpg image and update GDAL and MobilityDB, tweaking the build
scripting to bit a bit more efficient. Leave Apache Arrow v7.0.1 in
standard-cnpg, as dw-cnpg's use of tembo_parquet_s3_fdw and
parquet_s3_fdw from Trunk do not currently build with v19.

Use Trunk to install PostGIS, which is a dependency for MobilityDB, so
that we get the same binary as will be used in the stack.

Also change to the correct directory in the standard-cpnpg `Dockerfile`
to properly delete downloaded sources.