Install duckdb from Trunk pg_duckdb package (#82)
Required because the binary released by the DuckDB project does not work
with pg_duckdb, which always compiles its own.

While at it, upgrade to Trunk v0.15.10 and merge all the final commands
into a single `RUN` statement to cut down on the number of layers. Add a
TODO to remove all compilation in the final image to separate `FROM`
targets.