Upgrade to 17.3, 16.7, 15.11, and 14.16 (#89)
Refactor the generation of the postgres version matrix in
`.github/workflows/build_tembo_pg_slim.yaml` to include both the major
version and a full version with an underscore as used in Postgres tags.
Pass it to the `tembo-pg-slim/Dockerfile` as `PG_TAG` and use it to
perform a shallow clone of just the release tag. This way, any time
there are new versions, we can just edit the `TAGS` environment variable
to trigger an efficient rebuild with the latest version for all images.

While at it, give the jobs nicer names.