Create new Postgres image and bake pipeline (#96)
Create a new base image named `postgres` with a single `Dockerfile` and
a `docker-bake.hcl` file to script building it for various Postgres and
OS versions. This greatly simplifies things, as there is no need for
multiple `Dockerfile`s and base images. It's all managed via a single
call to `docker buildx bake` and various variables, with support for:

*   PostgreSQL 14-17
*   Ubuntu Noble and Jammy
*   ARM64 and AMD64

The tags include the Postgres version, OS name, and timestamp e.g.,

*   `postgres:17-jammy`
*   `postgres:17.4-jammy`
*   `postgres:17.4-jammy-202503041708`

Images built on the latest OS also have the tags:

*   `postgres:17`
*   `postgres:17.4`

And an image built on the latest Postgres includes the tag:

*   `postgres:latest`

The new workflow, `.github/workflows/bake.yaml`, takes advantage of the
recently-introduced `ubuntu-24.04-arm` GitHub worker to efficiently
build images for AMD64 and ARM64 in its matrixed jobs. It sets the
appropriate variables for each invocation of `docker buildx bake` to
build for each supported architecture, OS, and Postgres version. This
parallelized build process with native hardware provides the most
shortest possible build times.

To create multi-platform images, the `build` job pushes only SHAs, no
tags, records the `bake` metadata in files, and the `manifest` job
passes those files to `manifest.js`. This script assembles the necessary
parts to replicate the `docker-bake.hcl` tags to create and push
multi-platform manifests for each OS and Postgres version combination.

In general all version updates should be configured in the
`docker-bake.hcl` and `manifest.js` files, including the PostgreSQL
versions and Ubuntu versions and & releases.

The `bake.yaml` workflow runs on every push, but only pushes a release
on the `main` branch. A schedule configuration triggers new builds on
Mondays, so that we're always up-to-date.

The new image diverges from the old images in a number of significant
ways, some of which have implications for Trunk and the Tembo operator.
These changes were driven primarily by two goals: to better organize
persistent files on Tembo Cloud, and to transparently work in a plain
Docker container, as a CNPG cluster, and in Tembo Cloud. To that first
goal, the image now contains a simple `docker-entrypoint.sh` that
creates and starts a cluster, which should simplify testing in many
cases.

Other Details:

*   The image is about 1G in size, a quarter the size of the old slim
    image. This is mainly due to compiling Postgres in one stage with
    all the build tools (compiler, etc.), and copying the binaries to a
    separate stage that includes none of the build-time dependencies.
    This makes the image safer, as well.
*   The final image consists of a single layer. This reduces the size of
    the image, as well, and reduces the overhead on the Kubernetes
    cluster. This is done by copying all the files from the build stage
    onto a `FROM scratch` stage.
*   The image includes no additional extensions or Apt packages or
    compiled dependencies that extensions might require. The intention
    is that they will be installed by the operator into the persistent
    volume.
*   Postgres is installed into `/usr/lib/postgresql` instead of
    `/usr/lib/postgresql/$PG_MAJOR`. PGDG includes the version in the
    path so that one can install and run multiple versions at once, but
    we do not need that.
*   There is a new directory for the persistent volume,
    `/var/lib/postgresql/tembo`. The old directory,
    `/var/lib/postgresql/data`, still exists. This change allows the
    image to work with both Tembo Cloud and stock CNPG. However it will
    require logic in the Tembo operator to mount the proper directory.
*   The package directory (`pg_config --pkglibdir`) has moved from
    /var/lib/postgresql/data/tembo/$PG_MAJOR/lib` to
    `/usr/lib/postgresql/lib`, as the DSOs from core should always be
    for the current version, and don't need to persist. The location
    under `tembo`, which is the shared directory (`pg_config
    --sharedir`) was funky, as the shared directory is for
    architecturally-neutral files, and shared modules are not that.
*   A new directory, `/var/lib/postgresql/tembo/$PG_MAJOR/lib`, is
    intended for shared libraries from extensions. We will need to add
    this directory to `dynamic_library_path` and update `trunk` to
    install shared modules there instead of `pg_config --pkglibdir`.
*   Another new directory, `/var/lib/postgresql/tembo/$OS_NAME/lib`, is
    intended for system shared object libraries that dependencies
    require. We will need to build Apt packages to install them there.
*   A third new directory, `/var/lib/postgresql/tembo/$OS_NAME/etc`,
    contains the `ld.so.cache` file generated by `ldconfig`. Its
    canonical location, `/etc/ld.so.cache`, is a symlink to the new
    location. This will allow new shared files to be installed and
    `lcconfig` to update a persistent cache.
*   The `--datarootdir`, which is also used for what becomes `pg_config
    --sharedir`, now points to
    `/var/lib/postgresql/tembo/share`, and its contents are hard-linked
    in `/tmp/pg_sharedir`. The Tembo operator should be updated to always
    copy the latest data from `/tmp/pg_sharedir` on pod init, without
    wiping out any files or directories it doesn't copy.
*   The `postgres` user's home directory is now `/var/lib/postgresql`,
    and that is also the working directory, so `exec`ing into a shell
    starts in that directory, which has the `data` (for the default and
    CNPG clusters) and `tembo` (for Tembo Cloud) subdirectories
    immediately available.