Add custom `rpath` and `postgres-dev` image (#97)
Compile Postgres with a custom `rpath` that points not only to its main
lib directory, but also the lib directories under
`/var/lib/postgresql/tembo`. This will allow Postgres and extensions
compiled against this image to find third-party libraries in those
directories on the persistent volume without any changes to the
immutable parts of the image. In other words, no need for `ldconfig`.

As a consequence, remove the funky ld cache symlinking from the image,
as well as the ld configurations for the Postgres and Tembo lib
directories, as they're no longer needed.

In order to take advantage of this feature in future extension builds,
add a second image, `postgres-dev`, which contains `build-essential`,
`git`, `curl` and other tools for building extensions.

Update `manifest.json` to handle parse image names as part of the target
names.

In order to transparently support both the `postgres` and `postgres-dev`
images in `docker-bake.hcl`, remove all tagging and leave it to
`manifest.json`. This allows a single target with a matrix with the
target names, and a custom function to generate appropriate metadata
field values.