Add /tmp/sync-volume.sh (#101)
This script will be executed by the `initContainer` that Tembo Cloud
injects to run prior to CNPG's `initdb` container. Its main task is to
sync files from `/var/lib/postgresql/data` to a destination passed to
it, which will be the volume that will mask `/var/lib/postgresql/data`
in the CNPG containers. This ensures that all the files specific to this
version of PostgreSQL are up-to-date.

But its secondary task, defined in the `migrate()` function, is to
migrate files from the locations defined by the old Tembo images. It
deletes files and directories that are part of the PostgreSQL
`--libdir`, which are now included in this image, rather than stored on
the volume. But it leaves behind extension files, bitcode, and dependent
library files that are not part of core `--libdir`. This will allow
exiting images to transparently upgrade to the new image without loss of
functionality (although additional library dependencies still need to be
worked out).